cloud project/microservices/tts-service/requirements.txt
flask==3.0.0
kafka-python==2.0.2
boto3==1.34.0
gunicorn==21.2.0
